+++ core/DataTable/Manager.php
@@ -61,7 +61,7 @@ class Manager extends Singleton
     public function getTable($idTable)
     {
         if (!isset($this->tables[$idTable])) {
-            throw new TableNotFoundException(sprintf("This report has been reprocessed since your last click. To see this error less often, please increase the timeout value in seconds in Settings > General Settings. (error: id %s not found).", $idTable));
+            throw new TableNotFoundException(sprintf("Error: table id %s not found in memory. (If this error is causing you problems in production, please report it in Piwik issue tracker.)", $idTable));
         }
 
         return $this->tables[$idTable];