+++ plugins/Resolution/Columns/Configuration.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  *
  */
-namespace Piwik\Plugins\UserSettings\Columns;
+namespace Piwik\Plugins\Resolution\Columns;
 
 use Piwik\Columns\Dimension;
 use Piwik\Piwik;
@@ -15,6 +15,6 @@ class Configuration extends Dimension
 {
     public function getName()
     {
-        return Piwik::translate('UserSettings_ColumnConfiguration');
+        return Piwik::translate('Resolution_ColumnConfiguration');
     }
 }