+++ tests/PHPUnit/System/ImportLogsTest.php
@@ -103,3 +103,5 @@ class ImportLogsTest extends SystemTestCase
 }
 
 ImportLogsTest::$fixture = new ManySitesImportedLogs();
+ImportLogsTest::$fixture->includeIisWithCustom = true;
+ImportLogsTest::$fixture->includeNetscaler = true;